Some style changes
